pipeline/requirements.txt
feedparser>=6.0.0
newspaper3k>=0.2.8
# lxml_html_clean is installed separately in the Dockerfile
redis>=6.0.0
fastapi==0.95.0
uvicorn>=0.21.0
loguru>=0.7.0
schedule>=1.2.0
requests>=2.28.0
python-dateutil>=2.8.2
websockets>=10.0
sse-starlette>=1.6.1
